New FT for sharing, move session creation stuff to base
