+++ src/functional_tests/test_my_lists.py
@@ -3,36 +3,36 @@
 
 # Skipping the Login Process by Pre-creating a Session
 
-from django.conf import settings
+# from django.conf import settings
 from django.contrib.auth import get_user_model
 from selenium.webdriver.common.by import By
 from .base import FunctionalTest
-from .container_commands import create_session_on_server
-from .management.commands.create_session import create_pre_authenticated_session
+# from .container_commands import create_session_on_server
+# from .management.commands.create_session import create_pre_authenticated_session
 User = get_user_model()
 
 
 class MyListsTest(FunctionalTest):
-    def create_pre_authenticated_session(self, email):
-        # user = User.objects.create(email=email)
-        # session = SessionStore()
-        # session[SESSION_KEY] = user.pk 
-        # session[BACKEND_SESSION_KEY] = settings.AUTHENTICATION_BACKENDS[0]
-        # session.save()
-        if self.test_server:
-            session_key = create_session_on_server(self.test_server, email)
-        else:
-            session_key = create_pre_authenticated_session(email)
-        ## to set a cookie we need to first visit the domain.
-        ## 404 pages load the quickest!
-        self.browser.get(self.live_server_url + "/404_no_such_url/")
-        self.browser.add_cookie(
-            dict(
-                name=settings.SESSION_COOKIE_NAME,
-                value=session_key,  
-                path="/",
-            )
-        )
+    # def create_pre_authenticated_session(self, email):
+    #     # user = User.objects.create(email=email)
+    #     # session = SessionStore()
+    #     # session[SESSION_KEY] = user.pk 
+    #     # session[BACKEND_SESSION_KEY] = settings.AUTHENTICATION_BACKENDS[0]
+    #     # session.save()
+    #     if self.test_server:
+    #         session_key = create_session_on_server(self.test_server, email)
+    #     else:
+    #         session_key = create_pre_authenticated_session(email)
+    #     ## to set a cookie we need to first visit the domain.
+    #     ## 404 pages load the quickest!
+    #     self.browser.get(self.live_server_url + "/404_no_such_url/")
+    #     self.browser.add_cookie(
+    #         dict(
+    #             name=settings.SESSION_COOKIE_NAME,
+    #             value=session_key,  
+    #             path="/",
+    #         )
+    #     )
     
     def test_logged_in_users_lists_are_saved_as_my_lists(self):
         # Edith is a logged-in user
